+++ clang-tools-extra/test/clang-tidy/checkers/bugprone/exception-escape-rethrow.cpp
@@ -15,6 +15,7 @@ void callsRethrowerNoexcept() noexcept {
 
 int throwsAndCallsRethrower() noexcept {
 // CHECK-MESSAGES: :[[@LINE-1]]:5: warning: an exception may be thrown in function 'throwsAndCallsRethrower' which should not throw exceptions
+// CHECK-MESSAGES: :[[@LINE+2]]:9: note: may throw 'int' here
     try {
         throw 1;
     } catch(...) {
@@ -24,6 +25,7 @@ int throwsAndCallsRethrower() noexcept {
 
 int throwsAndCallsCallsRethrower() noexcept {
 // CHECK-MESSAGES: :[[@LINE-1]]:5: warning: an exception may be thrown in function 'throwsAndCallsCallsRethrower' which should not throw exceptions
+// CHECK-MESSAGES: :[[@LINE+2]]:9: note: may throw 'int' here
     try {
         throw 1;
     } catch(...) {
@@ -33,4 +35,19 @@ int throwsAndCallsCallsRethrower() noexcept {
 
 void rethrowerNoexcept() noexcept {
     throw;
+}
+
+void throwInt() {
+  throw 5;
+}
+
+void rethrowInt() noexcept {
+// CHECK-MESSAGES: :[[@LINE-1]]:6: warning: an exception may be thrown in function 'rethrowInt' which should not throw exceptions [bugprone-exception-escape]
+// CHECK-MESSAGES: :[[@LINE-5]]:3: note: may throw 'int' here
+  try {
+    throwInt();
+  } catch(int) {
+    throw;
+  } catch(...) {
+  }
 }